% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunk_data}
\alias{chunk_data}
\title{Split vector into chunk-sized chunks}
\usage{
chunk_data(x, chunk_size)
}
\arguments{
\item{x}{vector to split}

\item{chunk_size}{chunk size}
}
\value{
list of vectors sized max chunk size
}
\description{
Split vector into chunk-sized chunks
}
