% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetching.R
\name{fetch_by_country}
\alias{fetch_by_country}
\title{Fetch trade flows by partner country}
\usage{
fetch_by_country(chunk, start_date = "2011", end_date = "2022", cmd = "S4")
}
\arguments{
\item{chunk}{vector of valid reporter codes you can get from \link[UMARcomtradeR]{get_reporter_codes}}

\item{start_date}{character value od start year}

\item{end_date}{character value od end year}
}
\value{
fetched data or NULL
}
\description{
Function to fetch records of goods imports and exports for each reporting
country and partner country combination (for the totals only).
Fetches annual data for both imports and exports of goods over a specified
period of time. The period can be max 12 years.
}
\details{
Due to API limits but even more so due to random internal server errors that
keep cropping up, this funciton should be used with small chunks of reporter
countries. Chunk sizes of 3 work quite reliably.

The funciton also attemts 3 retries in case of server errors, and will report
any chunks that were not completed in the end.
}
