% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetching.R
\name{fetch_full}
\alias{fetch_full}
\title{Fetch trade flows by partner country and commodity codes for single country}
\usage{
fetch_full(
  reporter = "SVN",
  chunk,
  start_date = "2011",
  end_date = "2022",
  cmd = "S4",
  agg_l = 3
)
}
\arguments{
\item{reporter}{character value of valid reporter codes you can get from \link[UMARcomtradeR]{get_reporter_codes}}

\item{start_date}{character value od start year}

\item{end_date}{character value od end year}
}
\value{
fetched data or NULL
}
\description{
Function to fetch records of goods imports and exports for a sinlge reporting
country and all partner country combination by commodity codes.
}
\details{
Fetches annual data for both imports and exports of goods over a specified
period of time. The period can be max 12 years.

Due to API limits but even more so due to random internal server errors that
keep cropping up, this funciton may not always complete.

The funciton also attemts 3 retries in case of server errors.
}
