% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{full_single_country}
\alias{full_single_country}
\title{Scripting fun for fetch_full}
\usage{
full_single_country(
  reporter = "SVN",
  start_date = "2011",
  end_date = "2022",
  cmd = "S4",
  agg_l = 3,
  partner_codes = NULL,
  chunk_size = 5
)
}
\arguments{
\item{reporter}{singele country}

\item{start_date}{start year}

\item{end_date}{end year}

\item{cmd}{classification type}

\item{agg_l}{aggregation level}
}
\value{
dataframe with response
}
\description{
Scripting fun for fetch_full
}
