% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_commodity_codes}
\alias{get_commodity_codes}
\title{Get valid commodity codes at certain aggregation}
\usage{
get_commodity_codes(classification = "S4", level = 3)
}
\arguments{
\item{classification}{e.g "HS", "H6", "S3", "S4..."}

\item{level}{numeric, defaults to 3}
}
\value{
vector of legal classificaiton codes
}
\description{
Extracts a vector of commodity codes from a specific classification e.g.
"S4", "HS".. at a level of aggregation - this is determined solely based
on the number of digits in the classificaiton code. Level 3 means it has
three digits.
}
