% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_partner_codes}
\alias{get_partner_codes}
\title{Get partner codes - countries in ISO format}
\usage{
get_partner_codes()
}
\value{
character vector of valid funs.
}
\description{
Helper function to get vector of valid  partner codes using
ISO3 which is what `comtradr` funcitons require.
}
