% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_reporter_codes}
\alias{get_reporter_codes}
\title{Get reporter codes - countries in ISO format}
\usage{
get_reporter_codes()
}
\value{
character vector of valid funs.
}
\description{
Helper function to get vector of valid reporter (or partner) codes using
ISO3 which is what `comtradr` funcitons require.
}
