% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_vs_ct.R
\name{mz_check_partnerCode}
\alias{mz_check_partnerCode}
\title{Replacemente for ct_check_partnerCode}
\usage{
mz_check_partnerCode(partner, update = FALSE, verbose = FALSE)
}
\arguments{
\item{partner}{code}

\item{update}{logical}

\item{verbose}{logical}
}
\value{
iso code
}
\description{
Replacemente for ct_check_partnerCode
}
