% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_vs_ct.R
\name{mz_get_data}
\alias{mz_get_data}
\title{Replacement for ct_get_data}
\usage{
mz_get_data(
  type = "goods",
  frequency = "A",
  commodity_classification = "HS",
  commodity_code = "TOTAL",
  flow_direction = "all",
  reporter = "all",
  partner = "World",
  start_date = NULL,
  end_date = NULL,
  process = TRUE,
  verbose = FALSE,
  primary_token = comtradr::get_primary_comtrade_key(),
  mode_of_transport = "TOTAL modes of transport",
  partner_2 = "World",
  customs_code = "C00",
  update = FALSE,
  ...
)
}
\arguments{
\item{type}{The type of returned trade data. A character value. Possible values are: 'goods' for trade in goods and 'services' for trade in services. The default value is 'goods'.}

\item{frequency}{The frequency of returned trade data. A character value. Possible values are: 'A' for annual data and 'M' for monthly data. The default value is 'A'.}

\item{commodity_classification}{The trade (IMTS) classification scheme. A character value. The only possible value is 'HS'. This is the default.}

\item{commodity_code}{The commodity code(s). A character vector. All possible values are provided in the `comtradr::ct_get_ref_table()` function. You should use the relevant value from the `id` column. The default value is 'TOTAL': the sum of all commodities.}

\item{flow_direction}{The direction of trade flows. A character vector. Possible values are: 'import' for imports, 'export' for exports, 're-import' for re-imports, 're-export' for re-exports, or 'all' for imports, exports, re-imports, and re-exports. The default value is 'all'.}

\item{reporter}{Reporter ISO3 code(s). A character vector. Possible values include the desired country's ISO3 code. A full list of these can be found in the `comtradr::country_codes` dataset. 'all' can be supplied to return values for all reporter countries that are not labelled as 'group' by the UN (e.g. ASEAN countries). The default value is 'all'.}

\item{partner}{Partner ISO3 code(s). A character vector. Possible values include the desired country's ISO3 code. A full list of these can be found in the `comtradr::country_codes` dataset. 'all' can be supplied to return values for all reporter countries that are not labelled as 'group' by the UN (e.g. ASEAN countries). The default value is 'World' which returns the trade with all partner countries as an aggregate.}

\item{start_date}{The start date of the query. A character value. Yearly values should be in the format: `yyyy`. Monthly values should be in the format: `yyyy-mm`.}

\item{end_date}{The end date of the query. A character value. Yearly values should be in the format: `yyyy`. Monthly values should be in the format: `yyyy-mm`. This can be a maximum of 12 years after the start date for the annual data or one year after the start date for monthly data.}

\item{process}{A logical value. If TRUE, returns a data.frame with the results. If FALSE, returns the raw httr2 request. Defaults to TRUE.}

\item{verbose}{A logical value. If TRUE, sends status updates to the console. If FALSE, runs functions quietly.}

\item{primary_token}{Your primary UN Comtrade API token. A character value. Default is to check in environment for stored token, if not passed through the `comtradr::set_primary_comtrade_key` function.}

\item{mode_of_transport}{The Mode of Transport is set to `TOTAL modes of transport`}

\item{partner_2}{This value is set as a default to `0`, which is most likely the most general value and also the default on the Comtrade website.}

\item{customs_code}{The customs code is set to the default of `C00` which is the default for TOTAL across all customs procedures.}

\item{update}{A logical value. If TRUE, will download the possibly updated reference tables from the UN.}

\item{...}{You can pass in further parameters to the API that will not be checked and passed on as query parameters as is.}
}
\value{
response from comtrade
}
\description{
Replacement for ct_get_data
}
