% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{reporter_by_partner_total}
\alias{reporter_by_partner_total}
\title{Scripting fun for fetch_full}
\usage{
reporter_by_partner_total(
  start_date = "2011",
  end_date = "2022",
  cmd = "S4",
  reporter_codes = NULL,
  chunk_size = 5
)
}
\arguments{
\item{start_date}{start year}

\item{end_date}{end year}

\item{cmd}{classification type}

\item{reporter_codes}{char vector, probably from \link[UMARcomtradeR]{get_reporter_codes}}

\item{chunk_size}{defaults to 5}
}
\value{
dataframe with response
}
\description{
Scripting fun for fetch_full
}
