% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{reporter_by_world_w_cassifications}
\alias{reporter_by_world_w_cassifications}
\title{Scripting fun for fetch_by_commodity}
\usage{
reporter_by_world_w_cassifications(
  start_date = "2011",
  end_date = "2022",
  cmd = "S4",
  agg_l = 3,
  chunks = NULL
)
}
\arguments{
\item{start_date}{start year}

\item{end_date}{end year}

\item{cmd}{classification type}

\item{agg_l}{aggregation level}

\item{chunks}{list of vectors of valid reporter codes you can get from
\link[UMARcomtradeR]{get_reporter_codes}}
}
\value{
dataframe with response
}
\description{
Scripting fun for fetch_by_commodity
}
